fix(chore): move types
